+++ CHANGELOG.md
@@ -7,10 +7,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 Unreleased
 ***********
 * [BREAKING] Adds AAA protection to PHT endpoints
+* Updated ODA to v11.1.0
 * Adds auto-panel creation `POST` endpoint based on cycle description
 * Adds separate `PUT` endpoint to update panels
 * Updates ska-aaa-authhelpers version to 1.1.0
 * Updates `POST` reviews endpoint to check if a review already exist under a different id before creating new review
+* Added POST `/pht/proposal-access/create` for pht
+* Added GET `/pht/proposal-access/user` for pht
+* Added GET `/pht/proposal-access/user/{prsl_id}` for pht
+* Added PUT `/pht/proposal-access/user/{access_id}` for pht
 * Adds module_view page and diagrams to documentation
 * Adds the functionality to create permission when a proposal is created 
 * Adds the functionality to check permission when a user gets a proposal by id